Added comments. Put test function in correct file
Created utils file for tests
