+++ src/tests/test_main_view_zoom_pane_integ.py
@@ -17,6 +17,8 @@ from PySide2.QtCore import *
 from PySide2.QtGui import *
 from PySide2.QtWidgets import *
 
+from .utils import are_pixels_close, are_qrects_close
+
 class TestMainViewZoomPaneIntegration(unittest.TestCase):
 
     def setUp(self):
@@ -25,49 +27,6 @@ class TestMainViewZoomPaneIntegration(unittest.TestCase):
     def tearDown(self):
         self.test_model.close_app()
         del self.test_model
-    
-    def are_pixels_close(self, pixel1, pixel2):
-        '''
-        Helper functions to determine if two pixels are close. Used for when scrolling
-        in zoom pane and center's don't exactly align.
-        '''
-        if isinstance(pixel1, (QPoint, QPointF)):
-            pixel1 = (pixel1.x(), pixel1.y())
-
-        if isinstance(pixel2, (QPoint, QPointF)):
-            pixel2 = (pixel2.x(), pixel2.y())
-
-        pixel1_diff = abs(pixel1[0]-pixel1[1])
-        pixel2_diff = abs(pixel2[0]-pixel2[1])
-
-        diff_similar = abs(pixel1_diff - pixel2_diff) <= 2 
-
-        epsilon = 2
-        return abs(pixel1[0]-pixel2[0]) <= epsilon and diff_similar
-
-    def are_qrects_close(self, qrect1: QRect, qrect2: QRect, epsilon=2) -> bool:
-        top_left_1 = qrect1.topLeft()
-        top_left_1 = (top_left_1.x(), top_left_1.y())
-
-        width_1 = qrect1.width()
-        height_1 = qrect1.height()
-
-        top_left_2 = qrect2.topLeft()
-        top_left_2 = (top_left_2.x(), top_left_2.y())
-
-        width_2 = qrect2.width()
-        height_2 = qrect2.height()
-
-        width_diff = abs(width_1-width_2)
-        height_diff = abs(height_1-height_2)
-
-        diff_similar = abs(width_diff-height_diff) <= epsilon
-
-        return top_left_1 == top_left_2 \
-            and diff_similar \
-            and width_diff <= epsilon \
-            and height_diff <= epsilon
-
 
     def test_click_mv_highlight_zp(self):
         '''
@@ -247,7 +206,7 @@ class TestMainViewZoomPaneIntegration(unittest.TestCase):
         center_pixel_zp = self.test_model.get_zoom_pane_center_raster_point()
         center_pixel_mv = self.test_model.get_main_view_rv_center_raster_coord((0, 0))
 
-        self.assertTrue(self.are_pixels_close(center_pixel_mv, center_pixel_zp))
+        self.assertTrue(are_pixels_close(center_pixel_mv, center_pixel_zp))
 
     def test_scroll_zp_move_mv(self):
         '''
@@ -297,7 +256,7 @@ class TestMainViewZoomPaneIntegration(unittest.TestCase):
         center_pixel_zp = self.test_model.get_zoom_pane_center_raster_point()
         center_pixel_mv = self.test_model.get_main_view_rv_center_raster_coord((0, 0))
 
-        self.assertTrue(self.are_pixels_close(center_pixel_zp, center_pixel_mv))
+        self.assertTrue(are_pixels_close(center_pixel_zp, center_pixel_mv))
 
     def test_not_linked_highlight_box(self):
         '''
@@ -346,7 +305,7 @@ class TestMainViewZoomPaneIntegration(unittest.TestCase):
 
         zp_region = self.test_model.get_zoom_pane_visible_region()
     
-        self.assertTrue(self.are_qrects_close(zp_region, rv_00_region))
+        self.assertTrue(are_qrects_close(zp_region, rv_00_region))
         self.assertTrue(rv_01_region == None)
         self.assertTrue(rv_10_region == None)
 
@@ -399,59 +358,9 @@ class TestMainViewZoomPaneIntegration(unittest.TestCase):
 
         zp_region = self.test_model.get_zoom_pane_visible_region()
 
-        self.assertTrue(self.are_qrects_close(zp_region, rv_00_region))
+        self.assertTrue(are_qrects_close(zp_region, rv_00_region))
         self.assertTrue(rv_00_region == rv_01_region)
         self.assertTrue(rv_00_region == rv_10_region)
-    
-    def test_cp_highlight_box(self):
-        '''
-        Ensures that the context pane's compatible highlights is
-        just the dataset shown in the context pane
-        '''
-        # Create first array
-        rows, cols, channels = 75, 75, 3
-        # Create a vertical gradient from 0 to 1: shape (50,1)
-        row_values = np.linspace(0, 1, rows).reshape(rows, 1)
-        # Tile the values horizontally to get a 50x50 array
-        impl = np.tile(row_values, (1, cols))
-        # Repeat the 2D array across 3 channels to get a 3x50x50 array
-        np_impl = np.repeat(impl[np.newaxis, :, :], channels, axis=0)
-
-        # Create second array
-        # Create 49 linearly spaced values from 0 to 0.75 and then append a 0
-        row_values = np.concatenate((np.linspace(0, 0.75, rows - 5), np.array([0, 0, 0, 0, 0]))).reshape(rows, 1)
-        impl2 = np.tile(row_values, (1, cols))
-        np_impl2 = np.repeat(impl2[np.newaxis, :, :], channels, axis=0)
-
-        # Create third array
-        # Start with an array of zeros (50x1)
-        row_values = np.zeros((rows, 1))
-        # Choose the row index corresponding to 75% of the height.
-        nonzero_index = int(0.75 * (rows - 1))
-        row_values[nonzero_index] = 0.75
-        impl3 = np.tile(row_values, (1, cols))
-        np_impl3 = np.repeat(impl3[np.newaxis, :, :], channels, axis=0)
-
-        self.test_model.set_main_view_layout((2, 2))
-
-        ds1 = self.test_model.load_dataset(np_impl)
-        ds2 = self.test_model.load_dataset(np_impl2)
-        ds3 = self.test_model.load_dataset(np_impl3)
-
-        self.test_model.click_main_view_zoom_in()
-        self.test_model.click_main_view_zoom_in()
-        self.test_model.click_main_view_zoom_in()
-        self.test_model.click_main_view_zoom_in()
-
-        self.test_model.set_context_pane_dataset(ds1.get_id())
-
-        visible_region_00 = self.test_model.get_main_view_rv_visible_region((0, 0))
-        highlight_region = self.test_model.context_pane._get_compatible_highlights(ds1.get_id())[0]
-
-        # For an unknown reason, when I run this test inside of pytest and outside,
-        # I get two different results. Outside of pytests the below regions are the same, but
-        # in pytest, one of the values is off by 6, hence epsilon=6
-        self.assertTrue(self.are_qrects_close(highlight_region, visible_region_00, epsilon=6))
 
 
 
@@ -502,6 +411,6 @@ if __name__ == '__main__':
     print(f"visible_region_00: {visible_region_00}")
     print(f"highlight: {highlight}")
 
-    print(tester.are_qrects_close(highlight, visible_region_00))
+    print(are_qrects_close(highlight, visible_region_00))
 
     test_model.app.exec_()
